pyproject.toml
[project]
name = "ants"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]

requires-python = ">=3.13,<3.14"
dependencies = [
    "numpy (>=2.4.2,<3.0.0)",
    "pygame (>=2.6.1,<3.0.0)",
    "scipy (>=1.17.0,<2.0.0)",
    "tqdm (>=4.67.3,<5.0.0)",
    "torch>=2.0.0",
    "stable-baselines3>=2.0.0",
    "tensorboard>=2.0.0",
    "rich>=14.3.3",
    "setuptools>=69.5.1",
]


# [build-system]
# requires = ["poetry-core>=2.0.0,<3.0.0"]
# build-backend = "poetry.core.masonry.api"
